requirements.txt
fastapi
uvicorn
pydantic
numpy
requests
python-dotenv